feat(code): add mroe scalable html delivering
I read the html as string . Then I added jsdom to parse the static
html and access the DOM. Then I inject my qrcode as a div into a
link I had planted in the html file. Vola... This was done to apply
css, however it is not working. Have to start learning css????